GH-516 Make DataNeedsManagementController use correct HTTP Verbs

+++ core/src/main/java/energy/eddie/core/dataneeds/DataNeedsManagementController.java
@@ -1,7 +1,11 @@
 package energy.eddie.core.dataneeds;
 
 import energy.eddie.api.agnostic.DataNeed;
+import energy.eddie.api.agnostic.exceptions.DataNeedNotFoundException;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.*;
@@ -13,25 +17,26 @@ import java.net.URI;
  * common CRUD operations for data needs as documented in the APIS.md file.
  */
 @RestController
-@RequestMapping("/management/data-needs")
+@RequestMapping(path = DataNeedsManagementController.BASE_PATH_KEY, produces = MediaType.APPLICATION_JSON_VALUE)
 @ConditionalOnProperty(value = "eddie.data-needs-config.data-need-source", havingValue = "DATABASE")
 public class DataNeedsManagementController {
-    public static final String URL_PREFIX = "/management/data-needs";
-
+    public static final String BASE_PATH_KEY = "${management.server.urlprefix}/data-needs";
+    public final String basePath;
     private final DataNeedsDbRepository dataNeedsDbRepository;
 
-    public DataNeedsManagementController(DataNeedsDbRepository dataNeedsDbRepository) {
+    public DataNeedsManagementController(DataNeedsDbRepository dataNeedsDbRepository, @Value(BASE_PATH_KEY) String basePath) {
         this.dataNeedsDbRepository = dataNeedsDbRepository;
+        this.basePath = basePath;
     }
 
-    @RequestMapping(method = {RequestMethod.POST, RequestMethod.PUT})
+    @PostMapping(produces = MediaType.TEXT_PLAIN_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> createDataNeed(@NonNull @RequestBody DataNeedEntity newDataNeed) {
         final var id = newDataNeed.id();
         if (dataNeedsDbRepository.existsById(id)) {
-            return ResponseEntity.badRequest().body("data need with id " + id + " already exists");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("data need with id " + id + " already exists");
         }
         dataNeedsDbRepository.save(newDataNeed);
-        return ResponseEntity.created(URI.create(URL_PREFIX + "/" + id)).build();
+        return ResponseEntity.created(URI.create(basePath + "/" + id)).build();
     }
 
     @GetMapping()
@@ -40,17 +45,18 @@ public class DataNeedsManagementController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<DataNeed> getDataNeed(@PathVariable String id) {
-        final var dataNeed = dataNeedsDbRepository.findById(id);
-        return dataNeed.isEmpty() ? ResponseEntity.notFound().build() : ResponseEntity.ok(dataNeed.get());
+    public ResponseEntity<DataNeed> getDataNeed(@PathVariable String id) throws DataNeedNotFoundException {
+        return dataNeedsDbRepository.findById(id)
+                .<ResponseEntity<DataNeed>>map(ResponseEntity::ok)
+                .orElseThrow(() -> new DataNeedNotFoundException(id));
     }
 
-    @PostMapping("/{id}")
-    public ResponseEntity<String> updateDataNeed(@PathVariable String id, @RequestBody DataNeedEntity dataNeed) {
+    @PutMapping(path = "/{id}", produces = MediaType.TEXT_PLAIN_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> updateDataNeed(@PathVariable String id, @RequestBody DataNeedEntity dataNeed) throws DataNeedNotFoundException {
         if (!dataNeed.id().equals(id)) {
             return ResponseEntity.badRequest().body("data need id in url does not match data need id in body");
         } else if (!dataNeedsDbRepository.existsById(id)) {
-            return ResponseEntity.notFound().build();
+            throw new DataNeedNotFoundException(id);
         } else {
             dataNeedsDbRepository.save(dataNeed);
             return ResponseEntity.ok().build();
@@ -58,8 +64,11 @@ public class DataNeedsManagementController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteDataNeed(@PathVariable String id) {
+    public ResponseEntity<String> deleteDataNeed(@PathVariable String id) throws DataNeedNotFoundException {
+        if (!dataNeedsDbRepository.existsById(id)) {
+            throw new DataNeedNotFoundException(id);
+        }
         dataNeedsDbRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 }
